+++ src/main/java/tech/autodirect/api/database/TableUsers.java
@@ -16,6 +16,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
+import org.springframework.http.HttpStatus;
+import org.springframework.web.server.ResponseStatusException;
 import tech.autodirect.api.interfaces.TableOffersInterface;
 import tech.autodirect.api.interfaces.TableUsersInterface;
 
@@ -46,28 +48,32 @@ public class TableUsers extends Table implements TableUsersInterface {
         double downPayment,
         double budgetMonthly
     ) throws SQLException, ClassNotFoundException {
-        if (!userExists(userId)) {
-            PreparedStatement stmt = this.dbConn.prepareStatement(
-                    "INSERT INTO " + this.schemaName + "." + this.tableName + " VALUES (?, ?, ?, ?, ?);"
-            );
-            stmt.setString(1, userId);
-            stmt.setInt(2, creditScore);
-            stmt.setBigDecimal(3, BigDecimal.valueOf(downPayment));
-            stmt.setBigDecimal(4, BigDecimal.valueOf(budgetMonthly));
-            stmt.setString(5, TableOffersInterface.createTableName(userId));
+        PreparedStatement stmt = this.dbConn.prepareStatement(
+                "INSERT INTO " + this.schemaName + "." + this.tableName + " VALUES (?, ?, ?, ?, ?);"
+        );
+        stmt.setString(1, userId);
+        stmt.setInt(2, creditScore);
+        stmt.setBigDecimal(3, BigDecimal.valueOf(downPayment));
+        stmt.setBigDecimal(4, BigDecimal.valueOf(budgetMonthly));
+        stmt.setString(5, TableOffersInterface.createTableName(userId));
 
-            // Create offers for this user
-            TableOffersInterface tableOffers = new TableOffers(dbName);
-            tableOffers.newTable(userId);
+        // Create offers for this user
+        TableOffersInterface tableOffers = new TableOffers(dbName);
+        tableOffers.newTable(userId);
 
-            // execute and close the above SQL statement
-            stmt.executeUpdate();
-            stmt.close();
-        }
+        // execute and close the above SQL statement
+        stmt.executeUpdate();
+        stmt.close();
     }
 
     @Override
     public Map<String, Object> getUserByID(String userId) throws SQLException {
+        if (!checkUserExists(userId)) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "user not found"
+            );
+        }
+
         // construct a prepared SQL statement selecting the specified user
         PreparedStatement stmt = this.dbConn.prepareStatement(
             "SELECT * FROM " + this.schemaName + "." + this.tableName + " WHERE user_id = ?;"
@@ -86,6 +92,12 @@ public class TableUsers extends Table implements TableUsersInterface {
 
     @Override
     public void removeUserByID(String userId) throws SQLException {
+        if (!checkUserExists(userId)) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, "user not found"
+            );
+        }
+
         // construct a prepared SQL statement selecting the specified user
         PreparedStatement stmt = this.dbConn.prepareStatement(
             "DELETE FROM " + this.schemaName + "." + this.tableName + " WHERE user_id = ?;"
@@ -98,7 +110,7 @@ public class TableUsers extends Table implements TableUsersInterface {
     }
 
     @Override
-    public boolean userExists(String userId) throws SQLException {
+    public boolean checkUserExists(String userId) throws SQLException {
         // construct a prepared SQL statement selecting the specified user
         PreparedStatement stmt = this.dbConn.prepareStatement(
             "SELECT 1 FROM " + this.schemaName + "." + this.tableName + " WHERE user_id = ?;"
